moved and refactored
